requirements.txt
Django==1.11.13
django-bootstrap4==0.0.6
django-extensions==2.0.7
django-filter==1.1.0
django-registration==2.4.1
djangorestframework==3.8.2
factory-boy==2.11.1
